Pushing remaining files from laptop
