--- My project/Assets/Scripts/CreatureLogic/Actions/ActionBase.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d1b085faa88c393448e050d1933f14a8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
