+++ My project/Assets/Scripts/CreatureLogic/Actions/ActionGraph.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5b6004aa80f69eaff9412b0e1557071e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
