+++ My project/Assets/Scripts/CreatureLogic/Actions/ActionNode.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bc544970973654a7098b969f44d081fc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
