--- My project/Assets/Scripts/CreatureLogic/Actions/ExperimentalActionI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 243f43380fd25e946a96f3b6b3e64cf2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
