+++ My project/Assets/Scripts/CreatureLogic/Actions/ExperimentalActionI.txt.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d0c0ded1fe670c321b9df9fa4dcd34ff
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
