--- My project/Assets/Scripts/CreatureLogic/Actions/RequestBreed.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 132fcecc452e1c948bf5889db4e41765
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
