--- My project/Assets/Scripts/CreatureLogic/Actions/Responses/BreedResponse.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9bfaba34020b7a54abaa3a82eae2076e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
