--- My project/Assets/Scripts/CreatureLogic/Actions/Responses/Response.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c397daa47ef7b3d41bc7a57b1b186647
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
