--- My project/Assets/Scripts/CreatureLogic/Actions/Responses/ResponseAccepter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 00e94b3037f275c4a9b0b444e1c6f086
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
