--- My project/Assets/Scripts/CreatureLogic/Actions/Responses/ResponseMethod.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bcaf37bd6eb705c4fbd5da69ca63f08a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
