+++ My project/Assets/Scripts/CreatureLogic/Actions/Wander.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0c87a6f040b824e0f9c1b19c4bea76b5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
