--- My project/Assets/Scripts/CreatureLogic/BaseCreature.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 53b43d13004b01d498a6deeee1853750
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
