+++ My project/UserSettings/EditorUserSettings.asset
@@ -6,10 +6,10 @@ EditorUserSettings:
   serializedVersion: 4
   m_ConfigSettings:
     RecentlyUsedSceneGuid-0:
-      value: 5108015551070f08540a5a2011730a44434f192f7f7e71687e701f6be6b2626d
+      value: 540457545c5350035858542046765a44474e4e7d7d7d27367a2c4862b5b1663a
       flags: 0
     RecentlyUsedSceneGuid-1:
-      value: 540457545c5350035858542046765a44474e4e7d7d7d27367a2c4862b5b1663a
+      value: 5108015551070f08540a5a2011730a44434f192f7f7e71687e701f6be6b2626d
       flags: 0
     RecentlyUsedSceneGuid-2:
       value: 5b520d0503545b0d0c0c0a2715770748154f4d2c7d7d7e627a7d4a35b4e1646a
